fix big 分析表导入

+++ backend/ESL/src/main/java/io/apj/modules/workBook/entity/WorkOperationsEntity.java
@@ -47,7 +47,7 @@ public class WorkOperationsEntity implements Serializable {
 	/**
 	 * 频度
 	 */
-	private Integer frequency;
+	private Float frequency;
 	/**
 	 * 时间值
 	 */
